团队修改

+++ src/main/java/com/jingluo/jingluo/service/impl/GroupServiceImpl.java
@@ -1,8 +1,13 @@
 package com.jingluo.jingluo.service.impl;
 
+
+import com.jingluo.jingluo.dto.commondto.Page;
 import com.jingluo.jingluo.dto.groupdto.GroupDto;
 import com.jingluo.jingluo.dto.groupdto.GroupStuId;
 import com.jingluo.jingluo.dto.groupdto.TransferManagerDto;
+import com.jingluo.jingluo.common.LoggerCommon;
+import com.jingluo.jingluo.controller.FileController;
+import com.jingluo.jingluo.dto.*;
 import com.jingluo.jingluo.entity.Class;
 import com.jingluo.jingluo.entity.Group;
 import com.jingluo.jingluo.entity.Student;
@@ -12,11 +17,13 @@ import com.jingluo.jingluo.mapper.GroupMapper;
 import com.jingluo.jingluo.mapper.StudentMapper;
 import com.jingluo.jingluo.service.GroupService;
 import com.jingluo.jingluo.utils.IdCode;
+import com.jingluo.jingluo.utils.TokenUtil;
 import com.jingluo.jingluo.vo.ResultInfo;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
+
 
 import java.util.Date;
 import java.util.List;
@@ -27,31 +34,37 @@ public class GroupServiceImpl implements GroupService {
 
     @Autowired
     private GroupMapper groupMapper;
-
     @Autowired
     private StudentMapper studentMapper;
-
     @Autowired
     private ClassMapper classMapper;
-
+    @Autowired
+    public FileController fileController;
     /**
      * 创建团队
-     *
      * @param groupDto
-     * @param stuCode
      * @return
      */
     @Override
-    public ResultInfo groupCreate(GroupDto groupDto, String stuCode) {
+    public ResultInfo groupCreate(GroupDto groupDto) {
 
         try {
+            String userToken = groupDto.getUserToken();
+            String userCode = TokenUtil.getUserCodeFormToken(userToken);
+            //校验token是否失效
+            if (TokenUtil.tokenValidate(userToken, userCode)) {
+                LoggerCommon.error("登录已失效，请重新登陆");
+                return ResultInfo.fail("登录已失效，请重新登陆");
+            }
+
             //获取学生个人信息
-            Student stu = studentMapper.selectByCode(stuCode);
+            Student stu = studentMapper.selectByCode(userCode);
             if (stu == null) {
                 return ResultInfo.fail("学生信息不能为空");
             }
             StudentGroup sg = new StudentGroup();
             Group group = new Group();
+
             //判断学生是否加入过团队
             if (stu.getGroupId() == null || stu.getGroupId() == 0) {
                 //团队id
@@ -59,13 +72,15 @@ public class GroupServiceImpl implements GroupService {
                 //团队名称
                 group.setGroupName(groupDto.getGroupName());
                 //团队简介
-                group.setGroupInfo(groupDto.getGroupInfe());
+                group.setGroupInfo(groupDto.getGroupInfo());
                 //创建来源
                 group.setCreateFrom(groupDto.getCreateFrom());
                 //所在班级
                 group.setClassId(stu.getClassId());
                 //添加管理员
                 group.setStudentCode(stu.getStudentCode());
+                //图片
+                group.setHeadImg(groupDto.getHeadImg());
                 //团队成员加 1
                 group.setStuNum(1);
                 //创建时间
@@ -103,16 +118,28 @@ public class GroupServiceImpl implements GroupService {
      */
     @Override
     public ResultInfo deleteGroup(GroupStuId groupStuId) {
-        int count = groupMapper.deleteGroup(groupStuId.getGroupId());
-        int count2 = groupMapper.deleteStudentGroup(groupStuId.getGroupId());
-        //删除学生信息里的团队id
-        Student student = studentMapper.selectByCode(groupStuId.getStudentCode());
-        student.setGroupId(0);
-        int count3 = studentMapper.update(student);
-        if (count == 1 && count2 == 1 && count3 == 1) {
-            return ResultInfo.success("删除成功");
-        } else {
-            return ResultInfo.fail("删除失败");
+        try {
+            String userToken = groupStuId.getUserToken();
+            String userCode = TokenUtil.getUserCodeFormToken(userToken);
+            //校验token是否失效
+            if (TokenUtil.tokenValidate(userToken, userCode)) {
+                LoggerCommon.error("登录已失效，请重新登陆");
+                return ResultInfo.fail("登录已失效，请重新登陆");
+            }
+            int count = groupMapper.deleteGroup(groupStuId.getGroupId());
+            int count2 = groupMapper.deleteStudentGroup(groupStuId.getGroupId());
+            //删除学生信息里的团队id
+            Student student = studentMapper.selectByCode(groupStuId.getStudentCode());
+            student.setGroupId(0);
+            int count3 = studentMapper.update(student);
+            if (count == 1 && count2 == 1 && count3 == 1) {
+                return ResultInfo.success("删除成功");
+            } else {
+                return ResultInfo.fail("删除失败");
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResultInfo.fail("操作失败");
         }
 
     }
@@ -126,6 +153,13 @@ public class GroupServiceImpl implements GroupService {
     @Override
     public ResultInfo deleteGroupMember(GroupStuId groupStuId) {
         try {
+            String userToken = groupStuId.getUserToken();
+            String userCode = TokenUtil.getUserCodeFormToken(userToken);
+            //校验token是否失效
+            if (TokenUtil.tokenValidate(userToken, userCode)) {
+                LoggerCommon.error("登录已失效，请重新登陆");
+                return ResultInfo.fail("登录已失效，请重新登陆");
+            }
             Group group = groupMapper.findGroupId(groupStuId.getGroupId());
             //成员减一
             group.setStuNum(group.getStuNum() - 1);
@@ -151,6 +185,14 @@ public class GroupServiceImpl implements GroupService {
     @Override
     public ResultInfo addGroupMember(GroupStuId groupStuId) {
         try {
+            String userToken = groupStuId.getUserToken();
+            String userCode = TokenUtil.getUserCodeFormToken(userToken);
+            //校验token是否失效
+            if (TokenUtil.tokenValidate(userToken, userCode)) {
+                LoggerCommon.error("登录已失效，请重新登陆");
+                return ResultInfo.fail("登录已失效，请重新登陆");
+            }
+
             Student student = studentMapper.selectByCode(groupStuId.getStudentCode());
             student.setGroupId(groupStuId.getGroupId());
             studentMapper.update(student);
@@ -178,33 +220,84 @@ public class GroupServiceImpl implements GroupService {
 
     /**
      * 通过 班级名称 团队名称查询 团队
-     *
-     * @param selectName
-     * @param type
+     * @param selectGroupDto
      * @return
      */
     @Override
-    public ResultInfo selectGroup(String selectName, String type) {
-        // 0 表示班级名称
-        if (type.equals("0")) {
-            Class aClass = classMapper.selectById(selectName);
-            List<Group> group = groupMapper.selectGroupByClassId(aClass.getClassId());
-            if (group != null) {
-                return ResultInfo.fail("查询成功", group);
-            } else {
-                return ResultInfo.fail("其输入正确的班级");
+    public ResultInfo selectGroup(SelectGroupDto selectGroupDto) {
+        try {
+            String userToken = selectGroupDto.getUserToken();
+            String userCode = TokenUtil.getUserCodeFormToken(userToken);
+            //校验token是否失效
+            if (TokenUtil.tokenValidate(userToken, userCode)) {
+                LoggerCommon.error("登录已失效，请重新登陆");
+                return ResultInfo.fail("登录已失效，请重新登陆");
             }
-        }
-        //1 表示团队名称
-        if (type.equals("1")) {
-            Group group = groupMapper.selectGroupByGroupName(selectName);
-            if (group != null) {
-                return ResultInfo.fail("查询成功", group);
-            } else {
-                return ResultInfo.fail("其输入正确的团队名称");
+            // 0 表示班级名称
+            if (selectGroupDto.getType().equals("0")) {
+                Class aClass = classMapper.selectById(selectGroupDto.getSelevtName());
+                List<Group> group = groupMapper.selectGroupByClassId(aClass.getClassId());
+                if (group != null) {
+
+                    //实现分页功能
+                    Page page = new Page();
+                    page.setCurrentPage(selectGroupDto.getPage());
+                    //每页的开始数
+                    page.setStar((page.getCurrentPage() - 1) * page.getPageSize() + 1);
+                    //list的大小
+                    int count = group.size();
+                    //设置总页数
+                    page.setTotalPage(count % 10 == 0 ? count / 10 : count / 10 + 1);
+                    //对list进行截取
+                    page.setDataList(group.subList(page.getStar()
+                            , count - page.getStar() > page.getPageSize() ? page.getStar() + page.getPageSize() : count));
+                    return ResultInfo.success("查询成功", page);
+                } else {
+                    return ResultInfo.fail("其输入正确的班级");
+                }
             }
+            //1 表示团队名称
+            if (selectGroupDto.getType().equals("1")) {
+                List<Group>  group = groupMapper.selectGroupByGroupName(selectGroupDto.getSelevtName());
+                if (group != null) {
+                    //实现分页功能
+                    Page page = new Page();
+                    page.setCurrentPage(selectGroupDto.getPage());
+                    //每页的开始数
+                    page.setStar((page.getCurrentPage() - 1) * page.getPageSize() + 1);
+                    //list的大小
+                    int count = group.size();
+                    //设置总页数
+                    page.setTotalPage(count % 10 == 0 ? count / 10 : count / 10 + 1);
+                    //对list进行截取
+                    page.setDataList(group.subList(page.getStar()
+                            , count - page.getStar() > page.getPageSize() ? page.getStar() + page.getPageSize() : count));
+                    return ResultInfo.success("查询成功", page);
+                } else {
+                    return ResultInfo.fail("其输入正确的团队名称");
+                }
+            }
+            //查询全部
+            if (selectGroupDto.getType().equals("2")){
+                List<Group> AllGroup = groupMapper.selectAllGroup();
+                Page page = new Page();
+                page.setCurrentPage(selectGroupDto.getPage());
+                //每页的开始数
+                page.setStar((page.getCurrentPage() - 1) * page.getPageSize() + 1);
+                //list的大小
+                int count = AllGroup.size();
+                //设置总页数
+                page.setTotalPage(count % 10 == 0 ? count / 10 : count / 10 + 1);
+                //对list进行截取
+                page.setDataList(AllGroup.subList(page.getStar()
+                        , count - page.getStar() > page.getPageSize() ? page.getStar() + page.getPageSize() : count));
+                return ResultInfo.success("查询成功", page);
+            }
+            return ResultInfo.fail("查询失败,请正确输入要搜索的名称");
+        } catch (Exception e) {
+            e.printStackTrace();
+            return  ResultInfo.fail("查询失败哦");
         }
-        return ResultInfo.fail("查询失败,请正确输入要搜索的名称");
     }
 
     /**
@@ -215,14 +308,28 @@ public class GroupServiceImpl implements GroupService {
      */
     @Override
     public ResultInfo updataGroup(GroupDto groupDto) {
-        Group group = new Group();
-        group.setGroupName(groupDto.getGroupName());
-        group.setGroupInfo(groupDto.getGroupInfe());
-        int count = groupMapper.updateByPrimaryKeySelective(group);
-        if (count == 1) {
-            return ResultInfo.success("修改成功");
-        } else {
-            return ResultInfo.fail("修改失败");
+        try {
+            String userToken = groupDto.getUserToken();
+            String userCode = TokenUtil.getUserCodeFormToken(userToken);
+            //校验token是否失效
+            if (TokenUtil.tokenValidate(userToken, userCode)) {
+                LoggerCommon.error("登录已失效，请重新登陆");
+                return ResultInfo.fail("登录已失效，请重新登陆");
+            }
+            Group group = new Group();
+            group.setGroupName(groupDto.getGroupName());
+            group.setGroupInfo(groupDto.getGroupInfo());
+            group.setHeadImg(groupDto.getHeadImg());
+            group.setCreateFrom(groupDto.getCreateFrom());
+            int count = groupMapper.updateByPrimaryKeySelective(group);
+            if (count == 1) {
+                return ResultInfo.success("修改成功");
+            } else {
+                return ResultInfo.fail("修改失败");
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            return  ResultInfo.fail("修改失败");
         }
     }
 
@@ -234,38 +341,49 @@ public class GroupServiceImpl implements GroupService {
     @Override
     public ResultInfo transferManager(TransferManagerDto transferManagerDto) {
 
-        //根据老管理员获取学生信息
-        Student oldStudent = studentMapper.selectByCode(transferManagerDto.getOldAdministratorCode());
-        //根据老管理员获取这个团队信息
-        Group group = groupMapper.findGroupId(oldStudent.getGroupId());
-        //根基新管理员code 获取这个学生信息
-        Student xinStudent = studentMapper.selectByCode(transferManagerDto.getXinAdministratorCode());
+        try {
+            String userToken = transferManagerDto.getUserToken();
+            String userCode = TokenUtil.getUserCodeFormToken(userToken);
+            //校验token是否失效
+            if (TokenUtil.tokenValidate(userToken, userCode)) {
+                LoggerCommon.error("登录已失效，请重新登陆");
+                return ResultInfo.fail("登录已失效，请重新登陆");
+            }
+            //根据老管理员获取学生信息
+            Student oldStudent = studentMapper.selectByCode(transferManagerDto.getOldAdministratorCode());
+            //根据老管理员获取这个团队信息
+            Group group = groupMapper.findGroupId(oldStudent.getGroupId());
+            //根基新管理员code 获取这个学生信息
+            Student xinStudent = studentMapper.selectByCode(transferManagerDto.getXinAdministratorCode());
+            //判断转让后是否退出
+            if (transferManagerDto.getType().equals("1")) {
+                //判断是否是管理员 如果是 && 新管理员是这个团队中的成员
+                if (group.getStudentCode().equals(transferManagerDto.getOldAdministratorCode()) && xinStudent.getGroupId().equals(group.getGroupId())) {
+                    group.setStudentCode(transferManagerDto.getXinAdministratorCode());
+                    groupMapper.updateByPrimaryKeySelective(group);
+                    GroupStuId groupStuId = new GroupStuId();
+                    groupStuId.setGroupId(group.getGroupId());
+                    groupStuId.setStudentId(oldStudent.getStudentId());
+                    groupStuId.setStudentCode(oldStudent.getStudentCode());
+                    deleteGroupMember(groupStuId);
+                } else {
+                    return ResultInfo.fail("不是管理员或新管理员不是团队成员,请正确操作");
+                }
+            } else {
+                //判断是否是管理员 如果是 && 新管理员是这个团队中的成员
+                if (group.getStudentCode().equals(transferManagerDto.getOldAdministratorCode()) && xinStudent.getGroupId().equals(group.getGroupId())) {
+                    group.setStudentCode(transferManagerDto.getXinAdministratorCode());
+                    groupMapper.updateByPrimaryKeySelective(group);
+                } else {
+                    return ResultInfo.fail("请正确操作");
+                }
+            }
 
-        //判断转让后是否退出
-        if (transferManagerDto.getType().equals("1")) {
-            //判断是否是管理员 如果是 && 新管理员是这个团队中的成员
-            if (group.getStudentCode().equals(transferManagerDto.getOldAdministratorCode()) && xinStudent.getGroupId().equals(group.getGroupId())) {
-                group.setStudentCode(transferManagerDto.getXinAdministratorCode());
-                groupMapper.updateByPrimaryKeySelective(group);
-                GroupStuId groupStuId = new GroupStuId();
-                groupStuId.setGroupId(group.getGroupId());
-                groupStuId.setStudentId(oldStudent.getStudentId());
-                groupStuId.setStudentCode(oldStudent.getStudentCode());
-                deleteGroupMember(groupStuId);
-            } else {
-                return ResultInfo.fail("不是管理员或新管理员不是团队成员,请正确操作");
-            }
-        } else {
-            //判断是否是管理员 如果是 && 新管理员是这个团队中的成员
-            if (group.getStudentCode().equals(transferManagerDto.getOldAdministratorCode()) && xinStudent.getGroupId().equals(group.getGroupId())) {
-                group.setStudentCode(transferManagerDto.getXinAdministratorCode());
-                groupMapper.updateByPrimaryKeySelective(group);
-            } else {
-                return ResultInfo.fail("请正确操作");
-            }
+            return ResultInfo.success("管理员转让成功");
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResultInfo.fail("操作失败");
         }
-
-        return ResultInfo.success("管理员转让成功");
     }
 
 }
